third_party/WebKit/LayoutTests/fast/hidpi/dpr-setting.html

<!DOCTYPE html>
<script>
    window.targetScaleFactor = 2;
</script>
<script src="resources/srcset-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
    addEventListener("load", function() {
        shouldBeTrue('matchMedia("screen and (min-resolution: 2dppx)").matches');
    }, false);
</script>
<div>This test passes if the DPR is set to 2x</div>
